+++ src/components/Quests.tsx
@@ -5,7 +5,7 @@ import { graphql, QueryRenderer } from 'react-relay';
 import env from '../enviroment';
 import {
   QuestsQuery,
-  QuestsQueryResponse,
+  QuestsQueryResponse
 } from './__generated__/QuestsQuery.graphql';
 
 /**
